+++ R/specify.R
@@ -71,23 +71,7 @@ specify <- function(x, formula, response = NULL, explanatory = NULL) {
   else
     attr(x, "explanatory_type") <- class(x[[as.character(attr(x, "explanatory"))]])
   
-  # # NOT SURE if this is the best place for this?
-  # # Keep track of Satterthwaite degrees of freedom since lost when in aggregation w/
-  # # calculate()/generate()
-  # if (!is.null(attr(x, "response")) & 
-  #     !is.null(attr(x, "explanatory")) & 
-  #     !is.null(attr(x, "response_type")) & 
-  #     !is.null(attr(x, "explanatory_type"))){
-  #   if(attr(x, "response_type") %in% c("integer", "numeric") &
-  #      attr(x, "explanatory_type") == "factor"){
-  #     if(length(levels(x[[as.character(attr(x, "explanatory"))]])) == 2) {
-  #       attr(x, "distr_param") <- x %>% 
-  #         dplyr::summarize(stats::t.test(!! attr(x, "response") ~ !! attr(x, "explanatory"))[["parameter"]]) %>% 
-  #         dplyr::pull()
-  #     }
-  #   }
-  # }
-  
+  # Determine appropriate parameters for theoretical distribution fit
   x <- set_params(x)
   
   # add "infer" class