+++ NEWS.md
@@ -2,8 +2,12 @@
 - Updated documentation with examples
 - Created `pkgdown` site materials
 - Added `obs_stat` and `shade_direction` arguments to `visualize()`
-- Adding theoretical distributions to `visualize()`
-
+  - Not currently shading for `method = "both"`
+- Adding theoretical distributions to `visualize()` alone and as overlay
+  - Current implementations
+        - Two sample $t$
+        - ANOVA $F$
+        - One proportion $z$
 
 # infer 0.0.1
 - Implemented the "intro stats" examples for randomization methods as shown on the README